web/requirements.txt
flask
google-cloud-datastore==1.7.3	
requests==2.21.0	
wtforms==2.2.1	
pymongo==3.9.0	
dnspython==1.16.0
google-auth==1.6.3
